pyproject.toml
[project]
name = "shy-light"
version = "0.1.0"
description = "Camera → detection → OSC bridge for interactive 3D lighting in Unreal Engine"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.8",
    "mediapipe>=0.10",
    "deepface>=0.0.90",
    "python-osc>=1.8",
    "pyyaml>=6.0",
    "tf-keras>=2.20.1",
]

[project.optional-dependencies]
realsense = [
    "pyrealsense2>=2.54",
]

[project.scripts]
shy-light = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "main.py",
    "camera/**/*.py",
    "detectors/**/*.py",
    "state/**/*.py",
    "osc/**/*.py",
    "config.yaml",
]

[tool.hatch.build.targets.editable]
sources = ["."]
